TBC721: merge tbc721.test.ts into nft.test.ts as the classes NFT and TBC721 are in the same file
