+++ packages/TBC721/test/nft.test.ts
@@ -3,7 +3,8 @@
 import * as chai from 'chai'
 import { Computer } from '@bitcoin-computer/lib'
 import chaiMatchPattern from 'chai-match-pattern'
-import { NFT } from '../src/nft'
+import { NFT, TBC721 } from '../src/nft'
+
 import dotenv from 'dotenv'
 
 // If you want to connect to your local Bitcoin Computer Node, create a .env file 
@@ -89,6 +90,64 @@ describe('NFT', () => {
 
     it("Syncing to the NFT's revision returns an object that is equal to the NFT", async () => {
       expect(await receiver.sync(nft._rev)).deep.eq(nft)
+    })
+  })
+})
+
+describe('TBC721', () => {
+  const computer = new Computer({
+    url: 'http://127.0.0.1:1031',
+    network: 'regtest',
+  })
+  
+  let tbc721
+  let nft
+  
+  before(async () => {
+    await computer.faucet(1e7)
+  })
+  
+  describe('Constructor', () => {
+    it('Should create a new TBC721 object', async () => {
+      tbc721 = new TBC721(computer)
+      expect(tbc721.computer).deep.eq(computer)
+    })
+  })
+
+  describe('deploy', () => {
+    it('Should deploy a contract', async () => {
+      await tbc721.deploy()
+    })
+  })
+
+  describe('mint', () => {
+    it('Should mint an NFT', async () => {
+      nft = await tbc721.mint('name', 'symbol')
+    })
+  })
+
+  describe('balanceOf', () => {
+    it('Should return the balance', async () => {
+      const balance = await tbc721.balanceOf(computer.getPublicKey())
+      expect(balance).to.be.greaterThanOrEqual(1)
+    })
+  })
+
+  describe('ownerOf', () => {
+    it('Should return the owner', async () => {
+      const owners = await tbc721.ownersOf(nft._id)
+      expect(owners).deep.eq([computer.getPublicKey()])
+    })
+  })
+
+  describe('transfer', () => {
+    it('Should transfer an NFT', async () => {
+      const computer2 = new Computer()
+      const nft2 = await tbc721.mint('name', 'symbol')
+      const publicKey2 = computer2.getPublicKey()
+      await tbc721.transfer(nft2._id, publicKey2)
+      const res = await tbc721.balanceOf(computer.getPublicKey())
+      expect(res).to.be.greaterThanOrEqual(1)
     })
   })
 })
